feat(draft): add bot difficulty selector to cube draft setup
Cube draft had no UI to set AI difficulty — the draft store's difficulty
index defaulted to Medium and could never be changed for the cube flow,
even though the value was already plumbed through startCubeDraft and the
launch URL to the AI loop controller.

Extract the set-draft segmented difficulty control into a shared
BotDifficultySelector building block (reads/writes the draft store) and
mount it in the solo cube branch of DraftPage, above CubeSetupPanel. The
control lives in the solo page rather than the shared CubeSetupPanel so it
does not appear on multiplayer human pods (DraftPodPage).

Localize the difficulty labels: render via the new
setSelector.difficultyLevels keys (existing translations copied into all 7
draft.json locales) instead of the previously hardcoded English strings,
sourcing the id list from the now-exported canonical DIFFICULTY_NAMES.
